+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/jss/compare/v1.16.2...HEAD)
 ------------------------------------------------------------------------
 ### Changed
+- Moves compiled static assets from /static/build/ to /static/
 - Updates `COPYRIGHT_EMAIL` default value
 
 [1.16.2](https://github.com/jshwi/jss/releases/tag/v1.16.2) - 2022-05-11